+++ o3d/import/cross/targz_processor_test.cc
@@ -59,7 +59,7 @@ class ArchiveTestClient : public ArchiveCallbackClient {
   explicit ArchiveTestClient() : file_count_(0), index_(0) {}
   // ArchiveCallbackClient methods
   virtual void ReceiveFileHeader(const ArchiveFileInfo &file_info);
-  virtual bool ReceiveFileData(MemoryReadStream *stream, int nbytes);
+  virtual bool ReceiveFileData(MemoryReadStream *stream, size_t nbytes);
 
   int GetFileCount() const { return file_count_; }
   size_t GetNumTotalBytesReceived() const { return index_; }
@@ -90,7 +90,8 @@ void ArchiveTestClient::ReceiveFileHeader(const ArchiveFileInfo &file_info) {
 }
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-bool ArchiveTestClient::ReceiveFileData(MemoryReadStream *stream, int nbytes) {
+bool ArchiveTestClient::ReceiveFileData(MemoryReadStream *stream,
+                                        size_t nbytes) {
   const char *p = reinterpret_cast<const char*>(
       stream->GetDirectMemoryPointer());
